requirements-dev.txt
asgiref==3.6.0
boto3==1.28.35
botocore==1.31.35
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.2.1
django-cors-headers==4.0.0
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
idna==3.4
jmespath==1.0.1
oauthlib==3.2.2
Pillow==10.0.0
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
s3transfer==0.6.2
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
static3==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
